test/ru/javawebinar/basejava/storage/AbstractArrayStorageTest.java
package ru.javawebinar.basejava.storage;

import org.junit.Assert;
import org.junit.Test;
import ru.javawebinar.basejava.exception.StorageException;
import ru.javawebinar.basejava.model.Resume;

import static org.junit.Assert.assertEquals;

public abstract class AbstractArrayStorageTest extends AbstractStorageTest {

    protected final Resume[] array = {
            new Resume("name_1"),
            new Resume("name_2"),
            new Resume("name_3")};

    public AbstractArrayStorageTest(Storage storage) {
        super(storage);
    }

    @Test(expected = StorageException.class)
    public void saveStorageExceptionTest() {
        storage.clear();
        try {
            for (int i = 0; i < AbstractArrayStorage.STORAGE_LIMIT; i++) {
                storage.save(new Resume("name" + i));
            }
        } catch (StorageException e) {
            Assert.fail();
        }
        storage.save(new Resume("name"));
    }

    @Test
    public void checkIndexTest() {
        assertEquals(0, getIndex("name_1"));
    }

}